Crate servo_geometry

Enums

DeviceIndependentPixel

A normalized “pixel” at the default resolution for the display.

Traits

MaxRect

Functions

au_rect_to_f32_rect

A helper function to convert a rect of Au pixels to a rect of f32 units.

convert_rect_to_css_pixel

A helper function to convert a Device rect to CSS pixels.

convert_size_to_css_pixel

A helper function to convert a Device size to CSS pixels.

f32_rect_to_au_rect

A helper function to convert a rect of f32 pixels to a rect of app units.

Type Aliases

DeviceIndependentBox2D

DeviceIndependentIntLength

DeviceIndependentIntPoint

DeviceIndependentIntRect

DeviceIndependentIntSideOffsets

DeviceIndependentIntSize

DeviceIndependentIntVector2D

DeviceIndependentPoint

DeviceIndependentRect

DeviceIndependentSize

DeviceIndependentVector2D

FastLayoutTransform

FramebufferUintLength