Crate servo_geometry

Enums

DeviceIndependentPixel

A normalized “pixel” at the default resolution for the display.

Traits

MaxRect

Functions

au_rect_to_f32_rect

A helper function to convert a rect of Au pixels to a rect of f32 units.

f32_rect_to_au_rect

A helper function to convert a rect of f32 pixels to a rect of app units.

Type Definitions

FramebufferUintLength